Update Mail Importer
Seperate file per class and Zend with namespace
